Add run script extensions for a Django project
A new generator script, which simply uses the existing `run` generator
script, but with "django=true".

The template for the `run` script now adds many new lines for running a
Django project. The main changes are tht there is now a `./run django`
command, and that `./run start` now runs Django instead of `yarn run
start`.