run: Be more space-proof
The run script breaks if the project directory's path contains spaces.  This patch fix only a partially and requires .docker-project to be initialized to avoid crashes, the project parameter string still requires some sanitizing if we really want to determine it from the `pwd`.

This patch is NOT tested, please review.

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>